refactor: Add mappers to GiniCaptureNetworkService objects
PP-1728
